+++ tests/sample_runs/O1v5.py
@@ -5,7 +5,7 @@ CIRCUIT_DIR = path.expanduser("~/dev/TestData/O1.v5")
 
 args = (
     path.join(CIRCUIT_DIR, "builderRecipeAllPathways.xml"),   # recipe_file
-    path.join(CIRCUIT_DIR, "circuit.mvd3"),                   # mvd_file
+    path.join(CIRCUIT_DIR, "circuit.mvd3"),                   # circuit_file
     path.join(CIRCUIT_DIR, "touches/touchesData.*.parquet"),  # touch files
     "--driver-memory 4G"
 )