+++ tests/sample_runs/test_31k.py
@@ -8,7 +8,7 @@ DATA_2M = "/gpfs/bbp.cscs.ch/scratch/gss/leite/circuit_2M"
 print("Path for touches: " + path.join(DATA_31K, "circuit_spark/*.parquet"))
 args = (
     path.join(BASE_DIR, "recipe/builderRecipeAllPathways.xml"),  # recipe_file
-    path.join(DATA_31K, "circuit.mvd3"),                         # mvd_file
+    path.join(DATA_31K, "circuit.mvd3"),                         # circuit_file
     path.join(BASE_DIR, "morphologies/h5"),                      # morpho_dir
     path.join(DATA_31K, "circuit_spark/*.parquet")               # touch_files
 )