+++ tests/sample_runs/test_functionalizer.py
@@ -4,7 +4,7 @@ from os import path
 BASE_DIR = path.join(path.dirname(__file__), "..", "circuit_1000n")
 args = (
     path.join(BASE_DIR, "builderRecipeAllPathways.xml"),     # recipe_file
-    path.join(BASE_DIR, "circuit.mvd3"),                     # mvd_file
+    path.join(BASE_DIR, "circuit.mvd3"),                     # circuit_file
     path.join(BASE_DIR, "morphologies/h5"),                  # morpho_dir
     path.join(BASE_DIR, "touches/*.parquet"),    # touch_files
 )