+++ tools/validation/touches_dump.py
@@ -53,7 +53,7 @@ touches_old = sm.read.parquet(args.old)
 touches_new = sm.read.parquet(args.new)
 
 neuron_data = NeuronDataSpark(cppneuron.MVD_Morpho_Loader(args.neurons, args.morphologies), "_mvd")
-neuron_data.load_mvd_neurons_morphologies()
+neuron_data.load_neurons_morphologies()
 neurons = neuron_data.neuronDF