+++ CHANGES.rst
@@ -2,12 +2,13 @@
 Changelog
 =========
 
-Version 0.14.4
+Version 0.15.0
 ==============
 
 Changes:
   - Add additional output columns to gap-junction runs. See FUNCZ-211_.
   - Fix executions for circuits with only one synapse class. See FUNCZ-218_.
+  - Add preliminary SONATA support. See FUNCZ-217_.
 
 Version 0.14.3
 ==============
@@ -169,4 +170,5 @@ First working version with 3 base filters:
 .. _FUNCZ-201: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-201
 .. _FUNCZ-209: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-209
 .. _FUNCZ-211: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-211
+.. _FUNCZ-217: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-217
 .. _FUNCZ-218: https://bbpteam.epfl.ch/project/issues/browse/FUNCZ-218